power: qpnp-fg: add support to limit the learned capacity
For capacity learning algorithm, there is a requirement to cap
the learned capacity based on the specified limits. When the
capacity learning cycle completes, validate the current learning
capacity to see if it falls within the specified max/min limits.
If it goes above the max limit or goes below the min limit, then
cap the learned capacity to one of those limits as specified.

Add support for this requirement through the following device
tree properties.
- qcom,cl-max-limit-deciperc
- qcom,cl-min-limit-deciperc

CRs-Fixed: 978333
Change-Id: I9e42f6a4fd1d2e0e24d2e2a02379685463717b78
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>