msm: mdss: Fix AD backlight configuration
When AD backlight attenuation feature is not enabled, the backlight
send to panel doesn't need to go through the backlight linearization
and inverse linearization.

Also set ad_bl_level to the current panel backlight level when AD
is enabled the first time and reset the ad_bl_level to 0 when AD
is disabled.

Change-Id: If96eccd817ecee6eae5fcc56fda29197b8b9f50d
Signed-off-by: Ping Li <pingli@codeaurora.org>